Moved some code for rebuilding connectivity table; almost working
