+++ analyses/strength_analysis.py
@@ -37,11 +37,6 @@ from multipatch_analysis.connection_strength import (
 )
 
 
-@db.default_session
-def list_experiments(session):
-    return session.query(db.Experiment).all()
-
-
 class ExperimentBrowser(pg.TreeWidget):
     def __init__(self):
         pg.TreeWidget.__init__(self)
@@ -94,153 +89,7 @@ class ExperimentBrowser(pg.TreeWidget):
         self.scrollToItem(item)
 
 
-def get_amps(session, pair, clamp_mode='ic', get_data=False):
-    """Select records from pulse_response_strength table
-    """
-    cols = [
-        PulseResponseStrength.id,
-        PulseResponseStrength.pos_amp,
-        PulseResponseStrength.neg_amp,
-        PulseResponseStrength.pos_dec_amp,
-        PulseResponseStrength.neg_dec_amp,
-        PulseResponseStrength.pos_dec_latency,
-        PulseResponseStrength.neg_dec_latency,
-        PulseResponseStrength.crosstalk,
-        db.PulseResponse.ex_qc_pass,
-        db.PulseResponse.in_qc_pass,
-        db.PatchClampRecording.clamp_mode,
-        db.StimPulse.pulse_number,
-        db.StimSpike.max_dvdt_time,
-        db.PulseResponse.start_time.label('response_start_time'),
-    ]
-    if get_data:
-        cols.append(db.PulseResponse.data)
 
-    q = session.query(*cols)
-    q = q.join(db.PulseResponse)
-    
-    q, pre_rec, post_rec = join_pulse_response_to_expt(q)
-    q = q.join(db.StimSpike)
-    q = q.add_columns(post_rec.start_time.label('rec_start_time'))
-        
-    filters = [
-        (pre_rec.electrode==pair.pre_cell.electrode,),
-        (post_rec.electrode==pair.post_cell.electrode,),
-        (db.PatchClampRecording.clamp_mode==clamp_mode,),
-        (db.PatchClampRecording.qc_pass==True,),
-    ]
-    for filter_args in filters:
-        q = q.filter(*filter_args)
-    
-    # should result in chronological order
-    q = q.order_by(db.PulseResponse.id)
-
-    df = pandas.read_sql_query(q.statement, q.session.bind)
-    recs = df.to_records()
-    return recs
-
-
-def get_baseline_amps_NO_ORM_VERSION(session, expt, dev, clamp_mode='ic'):
-    # Tested this against the ORM version below; no difference in performance.
-    query = """
-    select 
-        brs.id, brs.pos_amp, brs.neg_amp, brs.pos_dec_amp, brs.neg_dec_amp
-    from 
-        baseline_response_strength brs
-    join baseline on brs.baseline_id = baseline.id
-    join recording on baseline.recording_id = recording.id
-    join patch_clamp_recording on patch_clamp_recording.recording_id = recording.id
-    join sync_rec on recording.sync_rec_id = sync_rec.id
-    join experiment on sync_rec.experiment_id = experiment.id
-    where 
-        experiment.id={expt_id} and
-        recording.device_key={dev_key} and
-        patch_clamp_recording.clamp_mode='{clamp_mode}' and
-        patch_clamp_recording.baseline_potential<=-50e-3 and
-        patch_clamp_recording.baseline_current>-800e-12 and
-        patch_clamp_recording.baseline_current<800e-12
-    """.format(expt_id=expt.id, dev_key=dev, clamp_mode=clamp_mode)
-    df = pandas.read_sql(query, session.bind)
-    recs = df.to_records()
-    return recs
-
-
-def get_baseline_amps(session, pair, clamp_mode='ic', amps=None, get_data=True):
-    """Select records from baseline_response_strength table
-
-    If *amps* is given (output from get_amps), then baseline records will be selected from the same
-    sweeps as the responses.
-    """
-    cols = [
-        BaselineResponseStrength.id,
-        BaselineResponseStrength.pos_amp,
-        BaselineResponseStrength.neg_amp,
-        BaselineResponseStrength.pos_dec_amp,
-        BaselineResponseStrength.neg_dec_amp,
-        BaselineResponseStrength.pos_dec_latency,
-        BaselineResponseStrength.neg_dec_latency,
-        BaselineResponseStrength.crosstalk,
-        db.Baseline.ex_qc_pass,
-        db.Baseline.in_qc_pass,
-        db.PatchClampRecording.clamp_mode,
-        db.Recording.start_time.label('rec_start_time'),
-        db.Baseline.start_time.label('response_start_time'),
-    ]
-    if get_data:
-        cols.append(db.Baseline.data)
-        
-    q = session.query(*cols)
-    q = q.join(db.Baseline).join(db.Recording).join(db.PatchClampRecording).join(db.SyncRec).join(db.Experiment)
-    
-    filters = [
-        (db.Recording.electrode==pair.post_cell.electrode,),
-        (db.PatchClampRecording.clamp_mode==clamp_mode,),
-        (db.PatchClampRecording.qc_pass==True,),
-    ]
-    for filter_args in filters:
-        q = q.filter(*filter_args)
-    
-    # should result in chronological order
-    q = q.order_by(db.Recording.start_time)
-
-    # if amps is not None:
-    #     q = q.limit(len(amps))
-
-    df = pandas.read_sql_query(q.statement, q.session.bind)
-    recs = df.to_records()
-
-    if amps is not None:
-        # for each record returned from get_amps, return the nearest baseline record
-        mask = np.zeros(len(recs), dtype=bool)
-        amp_times = amps['rec_start_time'].astype(float)*1e-9 + amps['response_start_time']
-        base_times = recs['rec_start_time'].astype(float)*1e-9 + recs['response_start_time']
-        for i in range(len(amps)):
-            order = np.argsort(np.abs(base_times - amp_times[i]))
-            for j in order:
-                if mask[j]:
-                    continue
-                mask[j] = True
-                break
-        recs = recs[mask]
-
-    return recs
-
-
-def join_pulse_response_to_expt(query):
-    pre_rec = aliased(db.Recording)
-    post_rec = aliased(db.Recording)
-    joins = [
-        (post_rec, db.PulseResponse.recording),
-        (db.PatchClampRecording,),
-        (db.SyncRec,),
-        (db.Experiment,),
-        (db.StimPulse, db.PulseResponse.stim_pulse),
-        (pre_rec, db.StimPulse.recording),
-    ]
-    for join_args in joins:
-        query = query.join(*join_args)
-
-    return query, pre_rec, post_rec
 
 
 class ResponseStrengthPlots(pg.dockarea.DockArea):
